decompose `Header` into `Logo` and `TabInfo` components
