+++ .changeset/real-ants-give.md
@@ -1,0 +1,5 @@
+---
+'snable': patch
+---
+
+decompose `Header` into `Logo` and `TabInfo` components
